requirements.txt
aerich==0.7.2
python-dotenv==1.0.0
quart-auth==0.9.0
quart-schema==0.18.0
quart==0.19.4
tortoise-orm==0.20.0
bcrypt==4.1.2
pydantic[email]==2.5.2
google-auth-oauthlib==1.2.0
asyncpg==0.29.0
humanize==4.9.0
Markdown==3.5.1
redis[hiredis]==5.0.1
unique-names-generator==1.0.2
pylint==3.0.3